+++ app/services/articles/updater.rb
@@ -42,7 +42,7 @@ module Articles
       # remove related notifications if unpublished
       if article.saved_changes["published"] == [true, false]
         Notification.remove_all_by_action_without_delay(notifiable_ids: article.id, notifiable_type: "Article", action: "Published")
-        Notification.remove_all(notifiable_ids: article.comments.pluck(:id), notifiable_type: "Comment")
+        Notification.remove_all(notifiable_ids: article.comments.pluck(:id), notifiable_type: "Comment") if article.comments.exists?
       end
       # don't send only if article keeps being unpublished
       dispatch_event(article) if article.published || was_published
